Adjust status calculations
There was an issue where app status was being used to process
deployments, jobs, etc but the app status wasn't being persisted. This
change should give us the status updates we are looking forward without
the "act before persist" issues.
