+++ pkg/controller/appdefinition/depends.go
@@ -1,16 +1,11 @@
 package appdefinition
 
 import (
-	"fmt"
-	"slices"
 	"strconv"
-	"strings"
 
 	"github.com/acorn-io/baaah/pkg/apply"
-	"github.com/acorn-io/baaah/pkg/typed"
 	v1 "github.com/acorn-io/runtime/pkg/apis/internal.acorn.io/v1"
 	"github.com/acorn-io/runtime/pkg/labels"
-	"k8s.io/apimachinery/pkg/util/sets"
 )
 
 func getDependencyAnnotations(app *v1.AppInstance, containerOrJobName string, deps []v1.Dependency) map[string]string {
@@ -72,14 +67,6 @@ func getDependencyAnnotations(app *v1.AppInstance, containerOrJobName string, de
 		s := app.Status.AppStatus.Containers[containerOrJobName]
 		s.Dependencies = depStatus
 
-		if !s.Ready {
-			msg, blocked := isBlocked(s.Dependencies, s.ExpressionErrors)
-			if blocked {
-				s.State = "waiting"
-			}
-			s.TransitioningMessages = append(s.TransitioningMessages, msg...)
-		}
-
 		if app.Status.AppStatus.Containers == nil {
 			app.Status.AppStatus.Containers = map[string]v1.ContainerStatus{}
 		}
@@ -88,14 +75,6 @@ func getDependencyAnnotations(app *v1.AppInstance, containerOrJobName string, de
 		s := app.Status.AppStatus.Jobs[containerOrJobName]
 		s.Dependencies = depStatus
 
-		if !s.Ready {
-			msg, blocked := isBlocked(s.Dependencies, s.ExpressionErrors)
-			if blocked {
-				s.State = "waiting"
-			}
-			s.TransitioningMessages = append(s.TransitioningMessages, msg...)
-		}
-
 		if app.Status.AppStatus.Jobs == nil {
 			app.Status.AppStatus.Jobs = map[string]v1.JobStatus{}
 		}
@@ -103,36 +82,4 @@ func getDependencyAnnotations(app *v1.AppInstance, containerOrJobName string, de
 	}
 
 	return result
-}
-
-func isBlocked(dependencies map[string]v1.DependencyStatus, expressionErrors []v1.ExpressionError) (result []string, _ bool) {
-	groupedByTypeName := map[string][]string{}
-
-	for depName, dep := range dependencies {
-		var key string
-		if dep.Missing {
-			key = string(dep.DependencyType) + " to be created"
-		} else if !dep.Ready {
-			key = string(dep.DependencyType) + " to be ready"
-		} else {
-			continue
-		}
-		groupedByTypeName[key] = append(groupedByTypeName[key], depName)
-	}
-
-	for _, exprError := range expressionErrors {
-		if exprError.DependencyNotFound != nil && exprError.DependencyNotFound.SubKey == "" {
-			key := string(exprError.DependencyNotFound.DependencyType) + " to be created"
-			groupedByTypeName[key] = append(groupedByTypeName[key], exprError.DependencyNotFound.Name)
-		}
-	}
-
-	for _, key := range typed.SortedKeys(groupedByTypeName) {
-		values := sets.New(groupedByTypeName[key]...).UnsortedList()
-		slices.Sort(values)
-		msg := fmt.Sprintf("waiting for %s [%s]", key, strings.Join(values, ", "))
-		result = append(result, msg)
-	}
-
-	return result, len(result) > 0
 }
